Fixed exception delegation, added recovery for website timeout. Progress from 300 to 450+ products supported on 490. Still ignoring products with mutliple prices

+++ src/com/olivierboucher/ear/MySQLHelper.java
@@ -66,7 +66,7 @@ public class MySQLHelper {
 			 IN iprice_start DATE,
 			 IN iprice_end DATE
 			*/
-			rqst.setInt(1,p.getId());
+			rqst.setLong(1,p.getId());
 			rqst.setString(2, p.getDescription());
 			rqst.setString(3, p.getNote());
 			rqst.setString(4, p.getSize());
